+++ errors.js
@@ -1,28 +1,30 @@
 'use strict';
-module.exports = create => {
-    const PortSQL = create('portSQL');
+module.exports = ({defineError, getError, fetchErrors}) => {
+    if (!getError('portSQL')) {
+        const PortSQL = defineError('portSQL', null, 'Port SQL generic');
 
-    return {
-        sql: PortSQL,
-        noConnection: create('noConnection', PortSQL, 'No connection to SQL server'),
-        parserError: create('parserError', PortSQL, 'Sql parser error'),
-        notReady: create('notReady', PortSQL, 'The connection is not ready'),
-        missingProcess: create('missingProcess', PortSQL, 'Invalid or missing resultset processing mode'),
-        notImplemented: create('notImplemented', PortSQL, 'Specified resultset processing mode is not implemented yet'),
-        unexpectedType: create('unexpectedType', PortSQL, 'Unexpected parameter type'),
-        unexpectedColumnType: create('unexpectedColumnType', PortSQL, 'Unexpected column type'),
-        invalidView: create('invalidView', PortSQL, 'Invalid view, cannot refresh'),
-        invalidResultSetOrder: create('invalidResultSetOrder', PortSQL, 'Invalid resultset order'),
-        duplicateResultSetName: create('duplicateResultSetName', PortSQL, 'Duplicate resultset name'),
-        singleResultExpected: create('singleResultExpected', PortSQL, 'Expected single or no result'),
-        wrongXmlFormat: create('wrongXmlFormat', PortSQL, 'Wrong XML format in result'),
-        retryFailedSchemas: create('retryFailedSchemas', PortSQL, 'Retries exceeded for failed schemas'),
-        noRowsExpected: create('noRowsExpected', PortSQL, 'No rows were expected in the result'),
-        singleResultsetExpected: create('singleResultsetExpected', PortSQL, 'Single resultset expected'),
-        oneRowExpected: create('oneRowExpected', PortSQL, 'Exactly one row was expected in the result'),
-        maxOneRowExpected: create('maxOneRowExpected', PortSQL, 'Maximum one row was expected in the result'),
-        minOneRowExpected: create('minOneRowExpected', PortSQL, 'Minimum one row was expected in the result'),
-        absolutePath: create('absolutePath', PortSQL, 'Absolute path error'),
-        invalidFileLocation: create('invalidFileLocation', PortSQL, 'Writing outside of base directory is forbidden')
-    };
+        defineError('noConnection', PortSQL, 'No connection to SQL server');
+        defineError('parserError', PortSQL, 'Sql parser error');
+        defineError('notReady', PortSQL, 'The connection is not ready');
+        defineError('missingProcess', PortSQL, 'Invalid or missing resultset processing mode');
+        defineError('notImplemented', PortSQL, 'Specified resultset processing mode is not implemented yet');
+        defineError('unexpectedType', PortSQL, 'Unexpected parameter type');
+        defineError('unexpectedColumnType', PortSQL, 'Unexpected column type');
+        defineError('invalidView', PortSQL, 'Invalid view, cannot refresh');
+        defineError('invalidResultSetOrder', PortSQL, 'Invalid resultset order');
+        defineError('duplicateResultSetName', PortSQL, 'Duplicate resultset name');
+        defineError('singleResultExpected', PortSQL, 'Expected single or no result');
+        defineError('wrongXmlFormat', PortSQL, 'Wrong XML format in result');
+        defineError('updateSchema', PortSQL, 'Schema update error');
+        defineError('retryFailedSchemas', PortSQL, 'Retries exceeded for failed schemas');
+        defineError('noRowsExpected', PortSQL, 'No rows were expected in the result');
+        defineError('singleResultsetExpected', PortSQL, 'Single resultset expected');
+        defineError('oneRowExpected', PortSQL, 'Exactly one row was expected in the result');
+        defineError('maxOneRowExpected', PortSQL, 'Maximum one row was expected in the result');
+        defineError('minOneRowExpected', PortSQL, 'Minimum one row was expected in the result');
+        defineError('absolutePath', PortSQL, 'Absolute path error');
+        defineError('invalidFileLocation', PortSQL, 'Writing outside of base directory is forbidden');
+    }
+
+    return fetchErrors('portSQL');
 };